requirements.txt
django-dramatiq>=0.9.1
APScheduler>=3.6.3
pydantic>=1.5.1
